requirements.txt
numpy
opencv-python
matplotlib
tqdm
spams-bin